+++ cypress/integration/Post.Images/I_add_all_required_fields.js
@@ -1,0 +1,8 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then("I add all required fields", () => {
+  cy.get('input[name="title"]')
+    .type('new post')
+    .get(".editor .ProseMirror")
+    .type('new post content')
+  })
