+++ cypress/integration/Post.Images/I_see_a_toaster_with_{string}.js
@@ -1,0 +1,5 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then("I see a toaster with {string}", (title) => {
+  cy.get(".iziToast-message").should("contain", title);
+})
