+++ cypress/integration/Post.Images/I_should_be_able_to_remove_the_image.js
@@ -1,0 +1,6 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then('I should be able to remove the image', () => {
+  cy.get('.delete-image-button' /* .dz-message > .base-button .crop-cancel*/)
+    .click()
+})
