+++ cypress/integration/Post.Images/my_post_has_a_teaser_image.js
@@ -1,0 +1,7 @@
+import { When } from "cypress-cucumber-preprocessor/steps";
+
+When('my post has a teaser image', () => {
+  cy.get('.contribution-form .image')
+    .should('exist')
+    .and('have.attr', 'src')
+})
