+++ cypress/integration/Post.Images/the_post_was_saved_successfully_with_the_{string}_teaser_image.js
@@ -1,0 +1,11 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then("the post was saved successfully with the {string} teaser image", condition => {
+  cy.get(".base-card > .title")
+    .should("contain", condition === 'updated' ? 'to be updated' : 'new post')
+    .get(".content")
+    .should("contain", condition === 'updated' ? 'successfully updated' : 'new post content')
+    .get('.post-page img')
+    .should("have.attr", "src")
+    .and("contains", condition === 'updated' ? 'humanconnection' : 'onourjourney')
+})
